local_ai_server/requirements_minimal.txt
# Minimal requirements for Local AI Server
# Only what we actually use - should be ~200MB instead of 8.78GB

# Core WebSocket communication
websockets==15.0.1

# Speech-to-Text (Vosk)
vosk==0.3.45

# Large Language Model (llama-cpp-python)
llama-cpp-python==0.3.16

# Data validation and config
pydantic==2.11.7
pydantic-settings==2.10.1
pyyaml==6.0.2

# Logging and utilities
structlog==25.4.0
tenacity==9.1.2

# Audio processing (minimal)
soundfile==0.13.1
numpy==1.26.4

# ARI client (for testing)
ari==0.1.3
swaggerpy==0.2.1
aiohttp==3.12.15

# Voice Activity Detection
webrtcvad-wheels==2.0.14
